requirements.txt
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.4
entrypoints==0.3
filelock==3.6.0
Flask==2.1.1
fonttools==4.28.5
gunicorn==19.9.0
idna==3.3
imageio==2.17.0
ipykernel==6.6.0
ipython==7.30.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mpmath==1.2.1
nest-asyncio==1.5.4
networkx==2.8
numpy==1.22.3
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.24
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.10.0
pyparsing==3.0.8
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pywin32==301; platform_system == "Windows"
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
sympy==1.10.1
tifffile==2022.4.8
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.1.1
windows-curses==2.3.0; platform_system == "Windows"
